Quick fix based on #79 to prevent unbounded loop unrolling
Context: array copies are called via a wrapper function - this took the
array length as a parameter. Now this is set directly (the parameter is 
kept and assigned to a dummy var to minimise code change as we might 
want to revert this in future. The dummy var is to prevent gnat error 
over unused var.